refactor: simplify updates slice.
