--- lib/features/updates/domain/domain.dart
@@ -1,5 +1,0 @@
-// Entities.
-export 'entities/update_data.dart';
-
-// Services.
-export 'services/watch_updates.dart';
