+++ lib/features/updates/models/update_entry.dart
