--- lib/features/updates/presentation/presentation.dart
@@ -1,12 +1,0 @@
-// Models.
-export 'models/update_entry.dart';
-
-// Pages.
-export 'pages/updates_page.dart';
-
-// Providers.
-export 'providers/updates_provider.dart';
-
-// Widgets.
-export 'widgets/chapter_update_tile.dart';
-export 'widgets/date_update_tile.dart';
